requirements.txt
Django==2.2.4
django-debug-toolbar==2.1
markdown2==2.3.8
pep8==1.7.1
pytz==2019.2
sqlparse==0.3.0
